+++ src/types/artifact/ArtifactData.ts
