+++ src/types/banner/BannerPhaseData.ts
