+++ src/types/banner/BannerRowData.ts
