+++ src/types/banner/ChronicledWishBannerPhaseData.ts
