+++ src/types/banner/ChronicledWishPhaseData.ts
