+++ src/types/character/CharacterOutfitData.ts
