+++ src/types/character/CharacterRowData.ts
