+++ src/types/character/CharacterStatsData.ts
