+++ src/types/character/CharacterStatsDataRow.ts
