+++ src/types/character/CharacterVAData.ts
