+++ src/types/tcg/TCGSkillData.ts
