+++ src/types/tcg/TCGSplashData.ts
