+++ src/types/tcg/TCGTalentsData.ts
