+++ src/types/weapon/WeaponRowData.ts
