+++ src/types/weapon/WeaponStatsData.ts
@@ -1,4 +1,4 @@
-import { SkillData } from "./SkillData"
+import { SkillData } from "../SkillData"
 
 export type WeaponStatsData = {
     atk: string,
