+++ src/types/weapon/WeaponStatsDataRow.ts
